+++ triggers/tests/BUILD.bazel
@@ -188,7 +188,7 @@ da_scala_library(
     da_scala_test_suite(
         name = "trigger-failure-integration-tests" + suffix,
         srcs = [
-            "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/FailureTesting.scala",
+            "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/LoadTesting.scala",
         ],
         data = [
             ":acs%s.dar" % suffix,
